requirements.txt
gradio
faiss-cpu
sentence-transformers
requests
groq